move APIs to V2 Resource
